services/ai/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
google-generativeai==0.7.2
requests==2.31.0
python-dotenv==1.0.1
